add shopping cart beta

+++ foodgram/api/v1/views.py
@@ -1,20 +1,22 @@
+from django.db.models import Avg
 from django.shortcuts import get_object_or_404
-from rest_framework import mixins, status, viewsets
+from rest_framework import mixins, serializers, status, viewsets
 from rest_framework.decorators import action
 from rest_framework.response import Response
 
 from head.models import (
-    Favorite,
     Ingredient,
     Recipe,
     Tag
 )
 from users.models import User
 from .serializers import (
-    FavoriteSerializer,
+    FavoriteShoppingSerializer,
+    FavoriteCreateSerializer,
     IngredientSerializer,
     RecipeSerializer,
     RecipeCreateSerializer,
+    ShoppingCreateSerializer,
     TagSerializer
 )
 
@@ -34,22 +36,71 @@ class RecipeViewSet(viewsets.ModelViewSet):
         serializer.save(author=user) # self.request.user
 
     @action(
+        methods=['get'], detail=False,
+    ) # permission_classes=(SelfEditUserOnlyPermission,)
+    def download_shopping_cart(self, request):
+        user = User.objects.get(id=1)  # user = User.objects.get(id=request.user)
+        cart = user.buy.values(
+            'recipe__ingredients__name',
+            'recipe__ingredients__measurement_unit'
+        ).annotate(total=Avg('recipe__ingredientrecipe__amount'))
+
+
+
+    @action(
         methods=['post', 'delete'], detail=True,
-        url_name='favorite',
     ) # permission_classes=(SelfEditUserOnlyPermission,)
     def favorite(self, request, pk):
-        user = User.objects.get(id=1)    # user = User.objects.get(id=request.user)
-        recipe = get_object_or_404(Recipe, id=pk)
+        user = User.objects.get(id=1) # user = User.objects.get(id=request.user)
+        recipe = self.get_object()
         if request.method == 'DELETE':
-            instance = user.favorites.get(recipe=recipe)
+            instance = user.favorites.filter(recipe=recipe)
+            if not instance:
+                raise serializers.ValidationError(
+                    {
+                        'errors': [
+                            'Этот рецепт в списке избранного отсутствует.'
+                        ]
+                    }
+                )
             instance.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
-        favorite = Favorite.objects.create(
-            recipe=recipe,
-            user=user
-        )
+        data = {
+            'user': user.id,
+            'recipe': pk
+        }
+        favorite = FavoriteCreateSerializer(data=data)
+        favorite.is_valid(raise_exception=True)
         favorite.save()
-        serializer = FavoriteSerializer(recipe)
+        serializer = FavoriteShoppingSerializer(recipe)
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+    @action(
+        methods=['post', 'delete'], detail=True,
+    ) # permission_classes=(SelfEditUserOnlyPermission,)
+    def shopping_cart(self, request, pk):
+        user = User.objects.get(id=1) # user = User.objects.get(id=request.user)
+        recipe = self.get_object()
+        if request.method == 'DELETE':
+            instance = user.buy.filter(recipe=recipe)
+            if not instance:
+                raise serializers.ValidationError(
+                    {
+                        'errors': [
+                            'Этот рецепт в списке покупок отсутствует.'
+                        ]
+                    }
+                )
+            instance.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        data = {
+            'user': user.id,
+            'recipe': pk
+        }
+        shop_cart = ShoppingCreateSerializer(data=data)
+        shop_cart.is_valid(raise_exception=True)
+        shop_cart.save()
+        serializer = FavoriteShoppingSerializer(recipe)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
